backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
google-genai
httpx
python-dotenv
